test of optional argument with small fix of OptionAttribute
